+++ MetaphysicsIndustries.Solus.Test/FunctionsT/Log10FunctionT/GetResultTest.cs
@@ -20,6 +20,7 @@
  *
  */
 
+using MetaphysicsIndustries.Solus.Functions;
 using NUnit.Framework;
 
 namespace MetaphysicsIndustries.Solus.Test.FunctionsT.Log10FunctionT
@@ -27,5 +28,26 @@ namespace MetaphysicsIndustries.Solus.Test.FunctionsT.Log10FunctionT
     [TestFixture]
     public class GetResultTest
     {
+        [Test]
+        public void ResultIsScalar()
+        {
+            // given
+            var arg1 = 1.ToNumber();
+            var args = new IMathObject[] { arg1 };
+            // precondition
+            Assert.IsTrue(arg1.IsScalar(null));
+            Assert.IsFalse(arg1.IsVector(null));
+            Assert.IsFalse(arg1.IsMatrix(null));
+            Assert.AreEqual(0, arg1.GetTensorRank(null));
+            Assert.IsFalse(arg1.IsString(null));
+            // when
+            var result = Log10Function.Value.GetResult(args);
+            // then
+            Assert.IsTrue(result.IsScalar(null));
+            Assert.IsFalse(result.IsVector(null));
+            Assert.IsFalse(result.IsMatrix(null));
+            Assert.AreEqual(0, result.GetTensorRank(null));
+            Assert.IsFalse(result.IsString(null));
+        }
     }
 }
