+++ MetaphysicsIndustries.Solus.Test/FunctionsT/Log10FunctionT/Log10FunctionTest.cs
@@ -20,8 +20,6 @@
  *
  */
 
-using MetaphysicsIndustries.Solus.Functions;
-using MetaphysicsIndustries.Solus.Values;
 using NUnit.Framework;
 
 namespace MetaphysicsIndustries.Solus.Test.FunctionsT.Log10FunctionT
@@ -29,26 +27,5 @@ namespace MetaphysicsIndustries.Solus.Test.FunctionsT.Log10FunctionT
     [TestFixture]
     public class Log10FunctionTest
     {
-        [Test]
-        public void ResultIsScalar()
-        {
-            // given
-            var arg1 = 1.ToNumber();
-            var args = new IMathObject[] { arg1 };
-            // precondition
-            Assert.IsTrue(arg1.IsScalar(null));
-            Assert.IsFalse(arg1.IsVector(null));
-            Assert.IsFalse(arg1.IsMatrix(null));
-            Assert.AreEqual(0, arg1.GetTensorRank(null));
-            Assert.IsFalse(arg1.IsString(null));
-            // when
-            var result = Log10Function.Value.GetResult(args);
-            // then
-            Assert.IsTrue(result.IsScalar(null));
-            Assert.IsFalse(result.IsVector(null));
-            Assert.IsFalse(result.IsMatrix(null));
-            Assert.AreEqual(0, result.GetTensorRank(null));
-            Assert.IsFalse(result.IsString(null));
-        }
     }
 }
